Refactoring Publish schools check controller
