+++ app/views/publish/providers/schools/checks/show.html.erb
